Moved to ImplUtils
